Extract StreamAVPlayerLoader into its own file
- Move the player-loading logic out of MediaViewer into
  Utils/AVPlayerLoader/StreamAVPlayerLoader.swift as a top-level type
- Rename the Utils/VideoCache folder to Utils/AVPlayerLoader
- Make VideoAttachmentCachingPolicy a final class with immutable properties
