+++ Sources/StreamChatSwiftUI/ChatMessageList/Gallery/MediaViewer.swift
@@ -353,7 +353,7 @@ struct StreamVideoPlayer: View {
     }
 
     private func loadPlayer() {
-        let loader = AVPlayerLoader(
+        let loader = StreamAVPlayerLoader(
             url: url,
             mediaLoader: utils.mediaLoader,
             avPlayerProvider: utils.avPlayerProvider,
@@ -369,96 +369,6 @@ struct StreamVideoPlayer: View {
                 avPlayer?.play()
             } catch {
                 self.error = error
-            }
-        }
-    }
-}
-
-extension StreamVideoPlayer {
-    @MainActor final class AVPlayerLoader {
-        private let url: URL
-        private let mediaLoader: MediaLoader
-        private let avPlayerProvider: AVPlayerProvider
-        private let cache: StreamVideoCache
-        private let policy: VideoAttachmentCachingPolicy
-        private let isPlayable: @Sendable (URL) async -> Bool
-
-        init(
-            url: URL,
-            mediaLoader: MediaLoader,
-            avPlayerProvider: AVPlayerProvider,
-            cache: StreamVideoCache,
-            policy: VideoAttachmentCachingPolicy,
-            isPlayable: @escaping @Sendable (URL) async -> Bool = AVPlayerLoader.isPlayable
-        ) {
-            self.url = url
-            self.mediaLoader = mediaLoader
-            self.avPlayerProvider = avPlayerProvider
-            self.cache = cache
-            self.policy = policy
-            self.isPlayable = isPlayable
-        }
-
-        func load() async throws -> AVPlayer {
-            let canCache = policy.maxCacheSize > 0 && !url.isFileURL && isContentTypeAllowed(url)
-            let key = url.path
-            let fileExtension = url.pathExtension.isEmpty ? "mp4" : url.pathExtension
-
-            guard canCache else {
-                return try await loadFromRemote()
-            }
-
-            if let localURL = await cache.completedFileURL(forKey: key, fileExtension: fileExtension) {
-                if await isPlayable(localURL) {
-                    return try await loadPlayer(from: MediaLoaderVideoAsset(asset: AVURLAsset(url: localURL)))
-                }
-                log.debug("Cached video is not playable; evicting and streaming from remote")
-                await cache.remove(forKey: key, fileExtension: fileExtension)
-            }
-
-            let fileRequest: MediaLoaderFileRequest
-            do {
-                fileRequest = try await mediaLoader.loadFileRequest(for: url)
-            } catch {
-                log.debug("Video cache file request failed; streaming without caching: \(error)")
-                return try await loadFromRemote()
-            }
-
-            let asset = StreamVideoAsset(
-                originalURL: url,
-                origin: fileRequest.urlRequest,
-                fileExtension: fileExtension,
-                cache: cache
-            )
-            return try await loadPlayer(from: MediaLoaderVideoAsset(asset: asset))
-        }
-
-        private func isContentTypeAllowed(_ url: URL) -> Bool {
-            guard let type = UTType(filenameExtension: url.pathExtension.lowercased()) else { return false }
-            return policy.allowedContentTypes.contains { type.conforms(to: $0) }
-        }
-
-        private nonisolated static func isPlayable(_ url: URL) async -> Bool {
-            await withCheckedContinuation { continuation in
-                let asset = AVURLAsset(url: url)
-                asset.loadValuesAsynchronously(forKeys: ["playable"]) {
-                    continuation.resume(
-                        returning: asset.statusOfValue(forKey: "playable", error: nil) == .loaded && asset.isPlayable
-                    )
-                }
-            }
-        }
-
-        private func loadFromRemote() async throws -> AVPlayer {
-            let videoAsset = try await mediaLoader.loadVideoAsset(at: url)
-            return try await loadPlayer(from: videoAsset)
-        }
-
-        private func loadPlayer(from videoAsset: MediaLoaderVideoAsset) async throws -> AVPlayer {
-            try await withCheckedThrowingContinuation { continuation in
-                avPlayerProvider.player(from: videoAsset) {
-                    continuation.resume(with: $0)
-                }
             }
         }
     }
