+++ Sources/StreamChatSwiftUI/Utils/AVPlayerLoader/StreamVideoAsset.swift
